# src/api/docs/openapi.yaml
openapi: 3.0.1
info:
  title: EPS API
  description: >-
    This is the API which supports the web and mobile applications for the
    Extended Prototype System (EPS).
  contact:
    email: jclements@ascend.network
  version: 0.29.0
...
paths:
...
  /surveys:
    get:
      tags:
        - surveys
      summary: Get list of surveys
      parameters:
        - in: query
          name: archived
          schema:
            type: string
          description: |
            This query parameter can be used to filter surveys by their
            `is_archived` property. It is not required and if not sent, the
            endpoint will return all non-archived surveys. If used, it supports
            three values: all, none, any. The result of using each filter values
            is:
              - all - returns only archived (`is_archived == true`) surveys
              - none - returns only non-archived (`is_archived == false`) surveys. The same behavior as not using the filter.
              - any - returns all surveys in the system
      description: |
        Returns array of surveys in the system. 
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfSurveys'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - surveys
      summary: Add survey to system.
      requestBody:
        description: Add new survey object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSurveyObject'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      security:
        - bearerAuth: [ ]
  /surveys/<int:survey_id>:
    get:
      tags:
        - surveys
      summary: Get a survey
      description: Returns survey based on URL
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - surveys
      summary: Update a survey
      description: |
        This endpoint will update a survey with the structure and name passed
        in.
      requestBody:
        description: |
          The endpoint accepts the following JSON object. The only
          required property is `id`. The endpoint will also verify that it
          matches the `survey_id` id the URL. If the `name` property is
          present, the endpoint will verify that it is not empty or null and
          that another survey does not have the same name.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSurveyObject'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Survey'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - surveys
      summary: Delete a survey
      description: |
        This endpoint will delete a survey and all of its responses from the
        system. The user submitting this request must have permission to
        delete surveys.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
  /surveys/<int:survey_id>/responses:
    get:
      tags:
        - surveys
      summary: Get list of responses for a survey
      parameters:
        - in: query
          name: archived
          schema:
            type: string
          description: >
            This query parameter can be used to filter survey responses by their

            `is_archived` property. It is not required and if not sent, the

            endpoint will return all non-archived survey responses. If used, it
            supports

            three values: all, none, any. The result of using each filter values

            is:
              - all - returns only archived (`is_archived == true`) survey responses
              - none - returns only non-archived (`is_archived == false`) survey responses. The same behavior as not using the filter.
              - any - returns all survey responses in the system
      description: Returns array of responses for a survey in the system
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfSurveyResponses'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - surveys
      summary: Add survey response
      description: This endpoint adds a new survey response to the system.
      requestBody:
        description: |
          The endpoint accepts a JSON object as defined below. The `structure`
          is required. The `status_id` is not. If a `status_id` is not sent or
          is empty, the status of the response will be set to the default
          response status. The endpoint will also validate that the
          `status_id` is valid.

          This endpoint also accepts location data. If location data is sent,
          it will be logged as the location the response was created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSurveyResponseObject'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyResponse'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
  /surveys/<int:survey_id>/responses/<int:response_id>:
    get:
      tags:
        - surveys
      summary: Get survey response
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyResponseWithSurvey'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - surveys
      summary: Update a survey response
      description: |
        This endpoint will update a survey response with the structure and
        status (if any) submitted.
      requestBody:
        description: |
          The endpoint accepts the following JSON object. `structure` is 
          required. `status_id` is not. If `status_id` is not provided, it
          will not be updated. The `status_id` is validated as a valid
          status id number.

          This endpoint also accepts location data. If location data is sent,
          it will be logged as the location the response was last updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSurveyResponseObject'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyResponse'
        '400':
          description: Invalid input
          content: { }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - surveys
      summary: Delete a survey response
      description: |
        This endpoint will delete a survey response from the system. The user
        submitting this request must have permission to delete surveys.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    ActionByUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        username:
          type: string
        name:
          type: string
        color:
          type: string
...
    Survey:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        structure:
          type: object
        is_archived:
          type: boolean
        responses:
          type: array
          description: Array of the system IDs for the responses for this survey.
          items:
            type: integer
        case_definitions:
          type: array
          description: Array of the system IDs for Case Definitions that use this survey.
          items:
            type: integer
        created_at:
          type: string
        updated_at:
          type: string
        created_by:
          #type: object
          $ref: '#/components/schemas/ActionByUser'
        updated_by:
          # type: object
          $ref: '#/components/schemas/ActionByUser'
    AddSurveyObject:
      type: object
      properties:
        name:
          type: string
        structure:
          type: object
        is_archived:
          type: boolean
      required:
        - name
        - structure
    UpdateSurveyObject:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        structure:
          type: object
        is_archived:
          type: boolean
      required:
        - id
    ArrayOfSurveys:
      type: array
      items:
        $ref: '#/components/schemas/Survey'
    SurveyResponse:
      type: object
      properties:
        id:
          type: integer
        survey:
          type: integer
        structure:
          type: object
        status:
          # type: object
          $ref: '#/components/schemas/SurveyResponseStatus'
        is_archived:
          type: boolean
        source_type:
          description: 'The source of the response, either Standalone or Case'
          type: string
        case:
          description: |
            The system ID of the Case this response is associated with if
            this response is for a Case.
          type: integer
        created_location:
          type: object
          description: The location the response record was created at
          $ref: '#/components/schemas/Location'
        updated_location:
          type: object
          description: The location the response record was last updated at
          $ref: '#/components/schemas/Location'
        created_at:
          type: string
        updated_at:
          type: string
        created_by:
          # type: object
          $ref: '#/components/schemas/ActionByUser'
        updated_by:
          # type: object
          $ref: '#/components/schemas/ActionByUser'
    SurveyResponseWithSurvey:
      type: object
      properties:
        id:
          type: integer
        survey_id:
          type: integer
        structure:
          type: object
        survey_structure:
          type: object
        status:
          # type: object
          $ref: '#/components/schemas/SurveyResponseStatus'
        is_archived:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
        created_by:
          type: object
          # $ref: '#/components/schemas/ActionByUser'
        updated_by:
          # type: object
          $ref: '#/components/schemas/ActionByUser'
    SurveyResponseStatus:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        default:
          type: boolean
      required:
        - id
        - name
    AddSurveyResponseObject:
      type: object
      properties:
        survey_id:
          type: integer
        structure:
          type: object
        status_id:
          type: integer
        is_archived:
          type: boolean
        source_type:
          description: 'The source of the response, either Standalone or Case'
          type: string
        latitude:
          type: number
          format: float
          description: Position's latitude in decimal degrees
        longitude:
          type: number
          format: float
          description: Position's longitude in decimal degrees
        position_accuracy:
          type: number
          format: float
          description: Accuracy of the latitude and longitude properties in meters
        altitude:
          type: number
          format: float
          description: Position's altitude in meters, relative to sea level
        altitude_accuracy:
          type: number
          format: float
          description: Accuracy of the altitude expressed in meters
        heading:
          type: number
          format: float
          description: The direction in which the device is traveling
        speed:
          type: number
          format: float
          description: The velocity of the device in meters per second
      required:
        - survey_id
        - structure
    UpdateSurveyResponseObject:
      type: object
      properties:
        id:
          type: integer
        survey:
          type: integer
        structure:
          type: object
        status_id:
          type: integer
        is_archived:
          type: boolean
        source_type:
          description: 'The source of the response, either Standalone or Case'
          type: string
        latitude:
          type: number
          format: float
          description: Position's latitude in decimal degrees
        longitude:
          type: number
          format: float
          description: Position's longitude in decimal degrees
        position_accuracy:
          type: number
          format: float
          description: Accuracy of the latitude and longitude properties in meters
        altitude:
          type: number
          format: float
          description: Position's altitude in meters, relative to sea level
        altitude_accuracy:
          type: number
          format: float
          description: Accuracy of the altitude expressed in meters
        heading:
          type: number
          format: float
          description: The direction in which the device is traveling
        speed:
          type: number
          format: float
          description: The velocity of the device in meters per second
      required:
        - structure
    ArrayOfSurveyResponses:
      type: array
      items:
        $ref: '#/components/schemas/SurveyResponse'
...
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: float
          description: Position's latitude in decimal degrees
        longitude:
          type: number
          format: float
          description: Position's longitude in decimal degrees
        position_accuracy:
          type: number
          format: float
          description: Accuracy of the latitude and longitude properties in meters
        altitude:
          type: number
          format: float
          description: Position's altitude in meters, relative to sea level
        altitude_accuracy:
          type: number
          format: float
          description: Accuracy of the altitude expressed in meters
        heading:
          type: number
          format: float
          description: The direction in which the device is traveling
        speed:
          type: number
          format: float
          description: The velocity of the device in meters per second
        location_recorded_dt:
          type: string
          format: date-time
          description: The timestamp the location was recorded
...
  responses:
    Success:
      description: Success
    UnauthorizedError:
      description: Unauthorized
    NotFoundError:
      description: NotFound
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT